Vm 64bit

+++ mType/ast/nodes/expressions/IntegerNode.hpp
@@ -6,13 +6,13 @@ namespace ast::nodes::expressions
     class IntegerNode : public ASTNode
     {
     private:
-        int value;
+        int64_t value;
 
     public:
-        explicit IntegerNode(int val, const SourceLocation& loc = SourceLocation());
+        explicit IntegerNode(int64_t val, const SourceLocation& loc = SourceLocation());
 
-        int getValue() const;
-        void setValue(int val);
+        int64_t getValue() const;
+        void setValue(int64_t val);
 
         Value accept(ASTVisitor<Value>& visitor) override;
         std::unique_ptr<ASTNode> clone() const override;
